backend/requirements.txt
flask==2.2.5
torch>=2.2.0
torchvision
Pillow
numpy==1.26.4

